Added links to InfoBox for day 547 | Додано посилання на InfoBox для дня 547
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Let's continue debunking myths about IT ARMY, based on the report by The Center for Strategic and International Studies: “Large-scale dissemination of DDoS tools to IT Army volunteers could give thousands of hackers the capacity to carry out cyberattacks, with the potential to use their skills for more nefarious purposes, a problem that governments have struggled with previously." For almost a year now, nearly all the software that we recommend to our subscribers has a closed configuration. This means that using it for purposes not associated with IT ARMY is not feasible.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Щоб не відставати від цілей, ми збираємося одночасно використовувати цілі db1000n із їхнього сховища GitHub. Ці оновлення відбуваються щодня, тож якщо db1000n зміниться, ймовірно, пройде ще кілька років, перш ніж зміни відобразяться тут. Повідомлення оприлюднило ІТ-Армія України:
Продовжимо розбирати міфи про IT ARMY, базуючись на звіті The Center for Strategic and International Studies: "Масове поширення інструментів DDoS серед добровольців IT ARMY може надати тисячам хакерів можливість проводити кібератаки, з потенційною здатністю використовувати свої навички для більш шкідливих цілей, проблемою з якою раніше вже боролись уряди." Вже рік як майже все програмне забезпечення, яке ми радимо нашим підписникам, має закритий конфіг. Це означає, що використовувати його для цілей, не пов'язаних з IT ARMY, не вийде.