Re-Deploy for day 558 (GitHub Actions previously failed issue: 2824)
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Maybe a bit late, but what is DDoS (Distributed Denial of Service)? DDoS  attacks aim to overload a target website or service by flooding them with excessive traffic. How it works:
1️⃣ Target Selection
2️⃣ Botnet Deployment: Devices infected with malware (bots) are controlled to send requests.
3️⃣ Traffic Flood: Bots bombard the target with fake requests, making it difficult for legitimate users to access the service.
🌟 Effectiveness: it's hard to distinguish malicious traffic from real users, making defense tricky.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Можливо трошки запізно, але що таке DDoS (Distributed Denial of Service)? DDoS мають на меті перевантажити онлайн сервіс, засипаючи їх надмірним трафіком. Як працює:
[1] Вибір Цілі
[2] Розгортання Ботнету: Комп'ютери, інфіковані шкідливим ПЗ (боти), використовуються для відправки запитів.
[3] Потік Трафіку: Боти бомбардують ціль фальшивими запитами, ускладнюючи доступ реальних користувачів до сервіса.
🌟 Чому це працює: важко відрізнити шкідливий трафік від трафіку реальних користувачів.