Added Day 567 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
The IT Army of Ukraine has indeed caused significant trouble for Russian bankers. Deputy Chairman of Sberbank's Management Board, Stanislav Kuznetsov, in an interview with TASS on September 10, complains that the activity of the so-called cybercriminals has broken all records. In particular, Kuznetsov notes: "There have been attacks from the so-called 'IT Army of Ukraine,' including, unfortunately, successful ones on the resources of regional government bodies of the Russian Federation, commercial banks, and insurance companies." Who's doing a great job? We are!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
IT Army of Ukraine таки допекла російським банкірам. Заступник голови правління Сбербанку Станіслав Кузнєцов в інтерв’ю ТАСС 10 вересня, жаліється на те, що активність так званих кіберзлочинців б’є всі рекорди. Зокрема, Кузнєцов зазначає: «З боку так званої "IT-Армії України" відбулися атаки, у тому числі, на жаль, результативні на ресурси регіональних органів державної влади РФ, комерційні банки та страхові компанії». Хто молодці? Ми молодці!