Додано день 569 - цілі, інформаційне поле та дні (TL) англ+укр
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Here's a quality counterpoint to this morning's article from The British Army Review. In it, you'll find a wealth of real insights into the vision and operation of IT ARMY.
"To mitigate risks in such a large organization, we've opted for a decentralized model of small groups whose tasks do not overlap, and we strive to keep their work as autonomous as possible. A volunteer who comes with an idea is more likely to get the opportunity to create their own group than to be added to an existing one."
Small group tactics aren't just for the battlefield, comrades.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Ловіть якісний антипод на ранкову статтю від The British Army Review. Тут ви знайдете величезну кількість реальних інсайдів про бачення і функціонування IT ARMY. 
"Для запобігання ризиків в такої великої організації, ми обрали розподілену модель маленьких груп, задачі яких не перетинаються і намагаємось тримати їх роботу як можна більш автономною. Волонтер який прийшов з ідеєю, скоріше отримає можливість створити власну групу ніж буде доданий до вже існуючої."
Тактика малих груп використовується не тільки на полі бою, комради.