Updated Info box for day 575 | Оновлено Інформаційне поле для дня 575
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Sometimes all it takes is a push in one place to set off a domino effect. Our operation yesterday against the RBS payment gateway broke Sberbank's acquiring system, which halted electronic payments in Pyatyorochka, Ural Airlines, and who knows where else. An incredible combo that we and our intelligence team can be proud of.
Join IT ARMY operations with the new ADSS solution for Linux, register in the bot, and become the drivers of cyber resistance.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Іноді достатньо натиснути в одному місці, щоб отримати ефект доміно. Наша вчорашня операція проти платіжного шлюзу РБС зламало еквайрінгову систему Сбербанку, що зупинило приймання електронних платежів в П'ятірці, Уральських Авіаліній і Бог ще знає кого. Неймовірне комбо, яким можемо пишатися ми й наша команда розвідки.
Приєднуйтесь до операцій IT ARMY з новим рішенням ADSS під Лінукс, реєструйтесь в боті та ставайте рушіями кіберспротиву.