Updated Info box for day 575
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
🚨 Cyber Warfare: A New Phenomenon in International Law 🚨
Karim Khan, the Prosecutor of the International Criminal Court, makes a clear statement: Cyber warfare is not just about bytes and bits; it's about real-world damage and suffering. He calls for the adaptation of international law to address this new form of conflict.
IT ARMY supports this idea and will steadfastly adhere to every letter of international legislation that will regulate cyber warfare.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
🚨 Кібервійна: Нове явище у міжнародному праві 🚨
Карім Хан, прокурор Міжнародного кримінального суду, робить однозначне твердження: кібервійна — це не просто байти та біти; це реальні збитки та страждання. Він закликає адаптувати міжнародне право для вирішення цієї нової форми конфлікту.
IT ARMY підтримує цю ідею і буде неухильно дотримуватись кожної літери міжнародного законодавства, що буде регулювати кібервійну.