Target Sync for day 576 | Цільова синхронізація на день 576
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
As a result of our cyber raid yesterday, internet providers, local propaganda media, and websites of the occupying administrations in Crimea were not functioning. Drones in the sky, IT ARMY in the wires—this is the picture of modern warfare. The amount of rude comments about the non-working services is astonishing. Is this the culture of the newcomers?
Join IT ARMY operations with the new ADSS solution for Linux, register in the bot, and become the drivers of cyber resistance.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
В результаті нашого вчорашнього кіберрейду в Криму не працювали інтернет-провайдери, місцеві пропагандистські ЗМІ, вебсайти окупаційних адміністрацій. Дрони в небі, IT ARMY в дротах - малюнок сучасної війни. Неодмінно дивує кількість хамських коментарів до непрацюючих сервісів. Це вже культура понаїхавших?
Приєднуйтесь до операцій IT ARMY з новим рішенням ADSS під Лінукс, реєструйтесь в боті та ставайте рушіями кіберспротиву.