Added Day 577 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Ukraine serves as a barrier against Russia's self-will in cyberspace, according to the head of the SBU Cybersecurity Department, Ilya Vityuk. In an interview with the BBC, published on September 6th, he noted, "I always say that Ukraine has debunked the myth of powerful Russian hackers." Moscow is throwing almost all of its cyber-expertise against Ukraine and that leaves it with little capacity to attack Western targets. If Ukraine falls, Mr Vitiuk warns, then those attacks will be directed elsewhere.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Україна – це бар’єр проти свавілля Росії у кіберпросторі, – так вважає начальник департаменту кібербезпеки СБУ Ілля Вітюк. В інтерв’ю BBC, що було опубліковане 6 вересня, він зазначає: "Я завжди кажу, що Україна розвінчала міф про могутніх російських хакерів". Москва кидає проти України майже всі свої кіберсили, і це залишає мало можливостей для атак на західні цілі. Але якщо Україна впаде, попереджає Вітюк, тоді ці атаки будуть спрямовані деінде.