Updated Info box for day 577 | Оновлено Інформаційне поле для дня 577
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
We continue the genocide of enemy ISPs, now under the sights of SimStar. Each such operation, together with the successful rear strikes of our army, increases the war fatigue of the enemy population, and therefore brings our victory closer. Without economic and moral exhaustion, the enemy will remain dangerous for a long time, so get your devices out IT-lery, every bit counts.
Join IT ARMY operations with the new ADSS solution for Linux, register in the bot, and become the drivers of cyber resistance.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Продовжуємо геноцид ворожих інтернет-провайдерів, тепер під прицілом SimStar. Кожна така операція разом з успішними тиловими ударами нашої армії підвищує втому від війни ворожого населення, а значить наближає нашу перемогу. Без економічного і морального виснаження ворог залишиться ще довго небезпечним, тому діставайте девайси айтилерія, важливий кожен біт.
Приєднуйтесь до операцій IT ARMY з новим рішенням ADSS під Лінукс, реєструйтесь в боті та ставайте рушіями кіберспротиву.