Target sync for day 580 | Цільова синхронізація на день 580
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
📢 Time to announce the winner of the IT ARMY chevron contest! 🏆 The design by Dmytro D garnered the most votes. The community has spoken, and this design will now become our official chevron! 🎉
Congratulations to the winner and thanks to everyone who participated in voting. The design will be made accessible to everyone. The vector file will be pinned in our Telegram chat and later published on our website.
Thank you for your active participation and support! This is a huge step in creating symbols that unite us. 
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
📢 Час оголосити переможця конкурсу шевронів IT ARMY! 🏆 Найбільшу кількість голосів зібрав дизайн від Dmytro D. Спільнота вирішила, і тепер цей дизайн стане нашим офіційним шевроном! 🎉
Вітаємо переможця і всіх, хто взяв участь у голосуванні. Дизайн буде доступний для всіх. Файл у векторі буде закріплений в нашому Telegram чаті та пізніше опублікований на нашому вебсайті.
Дякуємо за вашу активність і підтримку! Це великий крок у створенні символів, які об'єднують нас.