Updated info box for day 581 | Оновлено інформаційне поле для дня 581
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Subscribers of Win Mobile suspect that our previous victim, the internet provider "Volna," hired IT ARMY to fend off the competition. What about you? We're doing everything for free and for the best reasons.
By the way, we're offering a script to customer support services to explain why nothing is working—SMO, AAW, UFO, PSYOPS—all answers are equally valid.
Did you know that our statistics bot now shows your contribution to IT ARMY operations for the month, and we will be publishing the winners on the channel? Catch your moment!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Абоненти Win Mobile підозрюють що наша попередня жертва інтернет-провайдер "Волна" найняв IT ARMY, щоб здихатись від конкурента. Ну що ви? Ми все безплатно, із кращих міркувань.
До речі пропонуємо скрипт службам підтримки для пояснення чому нічого не працює - СВО, ППО, НЛО, ІПСО - всі відповіді однаково вірні.
Чи знаєте ви, що тепер бот статистики тепер показує ваш внесок в операції IT ARMY за місяць, а переможців ми будемо публікувати на каналі? Ловіть свій момент!