Added Day 587 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Want another nonsense from Russian propagandists? Here you go! On September 25-26, many news portals in Russia published an article with a headline along the lines of: 'Hackers from Russia and Ukraine breached Sony servers.' Can Ukrainians even do anything together with Russians? The answer is obvious: no. Indeed, when you read the primary source of the news about the cyberattack on Sony, there is no mention of the hackers' nationality.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Хочете ще одну нісенітницю від російських пропагандистів? Ловіть! 25-26 вересня чимало новинних порталів Росії опублікували статтю з приблизно таким заголовком: «Хакери з Росії та України зламали сервери Sony». Чи можуть взагалі українці робити щось спільно з росіянами? Відповідь очевидна: ні. І дійсно, коли почитати першоджерело новини про кібератаку на Sony, то ніякої згадки про національність хакерів там немає.