Added Day 609 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
The IT ARMY of Ukraine model is becoming the gold standard for organizing volunteers in cyberspace during wartime. This is becoming the norm for conflicts where the enemy has an extensive IT infrastructure.
We must clarify that any similarities are coincidental, IT ARMY of Ukraine has not and does not intend to transfer any technologies or expertise to groups siding with undemocratic regimes. Modern societies must first learn to control and take responsibility for their own governance, rather than seeking enemies.

/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Модель IT ARMY of Ukraine стає зразковим способом організації волонтерів у кіберпросторі в умовах війни. Це вже стає нормою для конфліктів, коли ворог має розгалужену IT-інфраструктуру.
Мусимо заявити що збіги випадкові й IT ARMY of Ukraine не передавала і не збирається передавати жодних технологій чи досвіду групам що воюють на боці недемократичних режимів. Народи сучасності мають спочатку навчитися контролювати та нести відповідальність за власну владу, а не шукати зовнішніх ворогів.